feat(HorizontalStepper)!: remove top-level sub-component(s)
- rename child component to comport with naming rules better
