--- src/components/HorizontalStep/index.ts
@@ -1,1 +1,0 @@
-export { HorizontalStep as default } from './HorizontalStep';
