+++ src/components/HorizontalStepper/HorizontalStepper.module.css
@@ -25,3 +25,77 @@
   flex-grow: 1;
   min-width: var(--eds-size-2);
 }
+
+/*------------------------------------*\
+    # HORIZONTAL STEP
+\*------------------------------------*/
+
+/**
+ * Horizontal Step displays the step text and associated decorative icon.
+ */
+ .horizontal-step {
+  display: flex;
+  align-items: center;
+  gap: var(--eds-size-1);
+  /**
+   * Required to prevent overflowing outside container.
+   */
+  min-width: 0;
+}
+
+/**
+ * Horizontal step variants
+ */
+.horizontal-step--complete {
+  color: var(--eds-theme-color-text-neutral-subtle);
+}
+.horizontal-step--active {
+  color: var(--eds-theme-color-text-neutral-strong);
+}
+.horizontal-step--incomplete {
+  color: var(--eds-theme-color-text-neutral-subtle);
+}
+
+/**
+ * Inherits the color from the parent.
+ *
+ * Changes icon from black to gray by default.
+ */
+.horizontal-step__number-icon {
+  color: inherit;
+}
+
+.horizontal-step__number-icon--incomplete {
+  --number-icon-incomplete-fill: var(--eds-theme-color-icon-disabled);
+}
+
+/**
+ * Horizontal Step Complete.
+ */
+.horizontal-step__complete-icon {
+  fill: var(--eds-theme-color-icon-utility-success);
+}
+
+/**
+ * Horizontal Step Incomplete.
+ */
+.horizontal-step__incomplete-icon {
+  margin-left: var(--eds-size-1);
+  margin-right: var(--eds-size-1);
+  fill: var(--eds-theme-color-icon-disabled);
+}
+
+/**
+ * Horizontal Step Text.
+ */
+.horizontal-step__text {
+  /* Hides text for smaller screen sizes. */
+  display: none;
+  @media all and (min-width: $eds-bp-md) {
+    display: inline;
+    /* Hides overflow text with ellipsis. */
+    white-space: nowrap;
+    overflow: hidden;
+    text-overflow: ellipsis;
+  }
+}
